# Add Negotiations mission (Red Alert, scg47ea)