--- a/mods/ra/maps/negotiations/notifications.yaml
+++ b/mods/ra/maps/negotiations/notifications.yaml
@@ -1,0 +1,6 @@
+Sounds:
+	Notifications:
+		HostageShot: gun27
+		DogWhine: dogw5
+		GuideOkay: guyokay1
+		Chronoshift: chrono2