--- a/mods/ra/missions.yaml
+++ b/mods/ra/missions.yaml
@@ -50,6 +50,7 @@ Aftermath Allied Missions:
 	in-the-nick-of-time
 	production-disruption
 	monster-tank-madness
+	negotiations
 Aftermath Soviet Missions:
 	shock-therapy
 	situation-critical