… mods/ra/maps/negotiations/map.bin
… mods/ra/maps/negotiations/map.png
… mods/ra/maps/negotiations/map.yaml
… mods/ra/maps/negotiations/negotiations.lua