This adds Negotiations from Aftermath. It's a short mission featuring Tanya, some hostages in need of rescue, and many barrels in need of exploding.

Normal's more of a straight conversion. Thoughts on the Hard changes are welcome.

Feedback on the cameras would be dandy. The original relies on vision of old areas to avoid patrols, so cameras are used to give information when needed.
 - Standard reveals are the usual 6s, while more event-y ones are 20s.
 - Most barrel explosions also come with a 1s camera reveal, for fun and confirming kills.

----

### Design changes:

**Original:** A hostage must die before the group will follow Tanya.
**New:** All hostages can be saved.

**Original:** If he lives, the helpful Guide will either follow Tanya until the mission ends, or stand still near the hostage fencing.
**New:** He can be escorted back to the church like the other civilians.

**Original:** Soviets produce tanks as soon as a War Factory is up, including Tesla Tanks.
**New:** Soviets will build a Service Depot before producing tanks. They also start with a Soviet Tech Center, which can be destroyed to stop production of Tesla Tanks or Shock Troopers.

**Original:** The General will retreat to the northwestern beach when Tanya or the Guide come within ~7 cells. A Soviet MCV will arrive if the General's left alone.
**New:** The General is faster and the retreat is activated sooner. Reinforcements will come if the General stays for ~~45s~~ -> 30s.
 - This is done because of ORA Tanya's better speed and range, and because of the added tank delay from the Service Depot.
 - The reinforcement timing change doesn't seem drastic. The MCV boat should arrive as Tanya and the hostages are ~~entering the forest~~ -> near the forest patrol's starting point.

The MCV's arrival is revealed, and there's a notification for the General's flare.
 
Swapped some unit types around for convention's sake.
 - Soviet Chinook reinforcements now enter from a Badger.
 - Aside from Tanya, Allied Badger reinforcements now enter from Chinooks.
 - The 3 Rocket Soldiers owned by the Soviets are swapped for Flamethrowers.

The Allied reinforcements come with 2 more mechanics.
 - May need adjusting. The attrition can be rough with a solo mechanic if Soviet reinforcements are signaled. The original solo mechanic is bonkers, healing a medium tank from near death in ~10s compared to ~64s in ORA.

----

### Hard difficulty:

All hostages must survive.

Tanya will encounter more early resistance.

 - If Tanya dies from a rocket, the launcher will be revealed (again) for the sake of clarity.

The return trip to the starting beach has some added enemies.

There is some more security around the prison and the BadGuy base area (where the MCV can deploy).

It's possible for Soviets to build Shock Troopers.

----

### Notes on porting and such:

To mimic [their original guard range](https://github.com/OpenRA/OpenRA/assets/4985264/fa691780-42da-4445-9400-b6d1d53d7103), some dogs have a wider AutoTarget ScanRadius.
 - I felt less kooky once I saw that allies-03a/allies-03b also did this for its dogs.
 - While the original rules.ini says `GuardRange=7`, 9 seemed to be a closer match.

Guide's house is more vulnerable to damage, since ORA rifles do little to structures.
 - The house will take ~11.56s to die if left alone. Original is ~10s.
 - Could have reduced the HP 53% -> 23%, but wanted to preserve the crumbling structure effect.
 
This western passage is originally passable, so I cleared a path. The eastern passage has a (narrow) path that works, so it was left untouched.
![RA95_RubbleWest](https://github.com/OpenRA/OpenRA/assets/4985264/941a5042-cab7-403a-baf6-795d95911aca)

The Demolition Truck's explosion range is tightened, closer to the original's.
 - The Soviet truck can otherwise explode barrels by the prison, freeing the hostages.

The Chrono Tank reinforcements' entry and triggers are slightly moved to avoid spawning inside other units.

- I don't know of a current way to order the Chrono Tanks' PortableChrono. I've instead implied that a Chronosphere is used to transport them.
  - This is consistent with Production Disruption or In The Nick of Time anyway.
  - They will be fully charged, and the Chronosphere sound is played instead of the tank shift sound.
  - I did manage the palette shift by moving a dummy unit. Mousetrap did something similar.

The patrolling tank encountered during the rescue is guarded by two Shock Troopers, so it's fortunately still a threat to ORA Tanya.

The Guide's movement is more direct.
- He no longer pauses at the southern river crossing since ORA Tanya has no issue keeping up.
- After the forest patrol is clear, he will head straight for the village bridge, instead of moving to the forest's middle or the forest patrol's starting point.

In case you're left to wonder why the Soviets' Demolition Truck goes hunting, that is the original `hunu` trigger.
 - BadGuys' units will hunt if Allied reinforcements have arrived and the General is dead.
 - This does not include units set to Area Guard. That leaves the truck, a handful of Flamethrowers, and BlockerTeam (1 Heavy Tank with 2 Shock Troopers).
 
For funsies, hostages following Tanya will panic if she dies. Their disregard for personal space in RA '95 meant they were likely to do it there as well.