@PunkPun 2023-07-26
> Swapped some unit types around for convention's sake.
>
> * Soviet Chinook reinforcements now enter from a Badger.
>* Aside from Tanya, Allied Badger reinforcements now enter from Chinooks.

I think this can be reversed. If anything we would like to give chinook to both factions like in vanilla