@PunkPun 2023-07-26
I think the flare next to the general can extinguish after he's been killed, maybe a few minutes death timer? It feels a bit off when it's permanent.

Also the camera on the demo truck should be triggered from the other side of the fence. I've also noticed that if you don't blow up the demo while on the other side of the fence, it's a big struggle to keep the civilians alive

I'm also a bit stumped, if you get into this situation how do you proceed?

![grafik](https://github.com/OpenRA/OpenRA/assets/37534529/768c27f7-a444-4480-92dc-b6e4e28b9fc7)