@JovialFeline 2023-07-26
> I think the flare next to the general can extinguish after he's been killed, maybe a few minutes death timer? It feels a bit off when it's permanent.

How long should that be? I know from fiddling with Don't Drink The Water that the original flares seem to expire after 8m but I'm guessing that'd be too long.

> Also the camera on the demo truck should be triggered from the other side of the fence. I've also noticed that if you don't blow up the demo while on the other side of the fence, it's a big struggle to keep the civilians alive

It can be, though in that case, you should have barrels for the village patrol; they move counter-clockwise and the dog especially should be easy to burn.

The demo truck camera should already spawn at the truck's death location unless it's attacking or I've goofed and this trigger only does it at its starting location.
```lua
	Trigger.OnKilled(SovietDemoTruck, function()
		if not SovietDemoTruck.IsIdle then
			return
		end
		SpawnPlayerCamera(SovietDemoTruck.Location, DateTime.Seconds(1), "camera.paradrop")
	end)
```
> I'm also a bit stumped, if you get into this situation how do you proceed?

The intended route seems to be returning to the forest. Once there, the tank and shock troopers at the river will move through, north toward the village bridge. You can use the same hideout on the east edge to wait them out, then proceed to the church.

I'm a little irked to see them right at the river crossing instead of in the middle of the open church area, though. Odd.