@PunkPun 2023-07-26
> The demo truck camera should already spawn at the truck's death location

Yeah it spawns, I mean the camera revealing the barrels