@JovialFeline 2023-08-04
### Update

The tank-guarding Shock Troopers are now Flamethrowers to reduce insta-deaths from fog.
  - At least for now, the forest team (with dogs) and northern Chinook team still have their Shock Troopers.

Allies are back to using Badgers for all their infantry drops.

Soviets use Chinooks and a handful of rocket launchers again.
  - There's some randomness in what northern edge cell the original Chinook spawns, but 9s until landing seems to be the average.

The explosion reveals should feel more responsive.

The money check for bot structure building is improved.

The General's flare will expire 30 seconds after he does. Tanya's flare expires after 5 minutes.

The timer now changes with difficulty.
  - Easy: 5 minutes
  - Normal: 4 minutes
  - Hard: 3 minutes
  - "every 5 minutes" is now "every few minutes" in the briefing text.

The "Keep all hostages alive" objective is still required for Hard, but also appears on Normal or Easy as an optional objective.

`negotiations` is listed with Tanya's usual targeting translation.

The copyright notice is added.