@PunkPun 2023-08-05
It's definitely easier with less shokies and with working patrols. Just the tiny cameras felt like there were enabled for too short. It's a blink it and you'll miss it