@PunkPun 2023-08-05
Without reading the briefing it might be difficult to to tell from where to call in allied reinforcements. I'm not wether we need to fix it, but it's something worth thinking about