@JovialFeline 2023-08-06
Small changes:
- Objectives now use the common utility functions, which had slipped my mind.
- After hostages makes it inside the church, a looping beacon is created at the starting beach. It will stop looping after Tanya triggers reinforcements.