@JovialFeline 2023-09-02
Thank you. I can't figure how that re-kill may have happened without Tanya shooting the hostage, but that possibility was easy enough to test and the new check should handle any premature deaths.

---

- If a hostage scheduled for execution is already dead, a random survivor will be chosen instead. If every hostage is dead, execution is skipped.

- The escort and signal objectives are delayed until they're more relevant. The beach beacon is removed.

- The destroy objective is declared after the hostages are escorted. It's still possible to destroy all Soviets for victory without reinforcements, so long as hostages are escorted to the church at some point.

- A Chrono Tank timing goof is fixed. The tanks can be triggered after ~~hostages enter the church~~ -> Allied reinforcements are signaled.

- The hostage pen's executioner is set to Defend stance.

- The returning forest patrol will have another waypoint in their path so their dogs won't run ahead as much, and another reveal is added. The reveals for that group's return can trigger after ~~hostages are freed~~ -> after the group moves all the way south to the Guide house, like Blocker Team does for their own reveal.

- Blocker Team's path to the Guide house skips the forest patrol's starting point and makes a traffic jam less likely.