@JovialFeline 2023-09-03
I've considered including that Tech Center in the fire sale if the General never signaled for help. It's owned by BadGuy (owner of the MCV). It should be easy enough to add once that check is done; `FireSale` does its thing with whatever player gets passed as an argument.

I'll have to ponder the objective stuff more. Ideally, I'd add the destroy objective after a delay but the game will see all objectives done and trigger victory. That is less of a problem on Hard where saving every hostage is required and only gets ticked done once all Soviets are gone.

Funny enough, I used to have SignalFlareNorth as the notification but went back and forth on it. Perhaps the west one will fit better with how the player's map is gonna look.

D'ya have any particular ideas for the church? Should I bring back the beacon?  :p